Remove circular imports. Add new jest and eslint configuration
