+++ package.json
@@ -7,7 +7,8 @@
     "build": "next build",
     "start": "next start",
     "test": "jest",
-    "check": "tsc --noEmit"
+    "check": "tsc --noEmit",
+    "circularity": "madge --circular --extensions ts,tsx ./src"
   },
   "dependencies": {
     "@lauf/lauf-store": "^0.1.0-beta.8",
@@ -30,9 +31,12 @@
     "@types/file-saver": "^2.0.2",
     "@types/jest": "^26.0.23",
     "@types/styled-components": "^5.1.7",
+    "@typescript-eslint/eslint-plugin": "^4.26.0",
+    "@typescript-eslint/parser": "^4.26.0",
+    "eslint": "^7.27.0",
     "jest": "^27.0.1",
     "madge": "^4.0.2",
     "ts-jest": "^27.0.1",
-    "typescript": "^4.2.2"
+    "typescript": "^4.3.2"
   }
 }
