+++ src/util.tsx
@@ -2,10 +2,16 @@ import React from "react";
 import { saveAs } from "file-saver";
 import { pdf } from "@react-pdf/renderer";
 import { Immutable, Store } from "@lauf/lauf-store";
-import { AppState, Category, Entry, ScoreName, Scorer, Tag } from "./types";
+import {
+  AppState,
+  CATEGORIES,
+  Category,
+  Entry,
+  ScoreName,
+  Scorer,
+  Tag,
+} from "./types";
 import { Resume } from "./components";
-import { SCORERS } from "./score";
-import { CATEGORIES } from "./data";
 
 export const LAUNCH_TIME = new Date().getTime();
 
@@ -54,3 +60,24 @@ export function createTagsScorer(...tags: Tag[]): Scorer {
     return 0;
   };
 }
+
+export const SCORERS: Record<ScoreName, Scorer> = {
+  // boost: (entry) => entry.boost || 0,
+  recency: (entry) => -(entry.stop ? LAUNCH_TIME - entry.stop.getTime() : 0), //Negative reverses order
+  duration: (entry) =>
+    entry.stop
+      ? entry.stop.getTime() - entry.start.getTime()
+      : LAUNCH_TIME - entry.start.getTime(),
+  employment: createTagsScorer("employment"),
+  education: createTagsScorer("education"),
+  society: createTagsScorer("society"),
+  coding: createTagsScorer("coding"),
+  electronics: createTagsScorer("electronics"),
+  management: createTagsScorer("management"),
+  "machine learning": createTagsScorer("machine learning"),
+  invention: createTagsScorer("invention"),
+  design: createTagsScorer("design"),
+  art: createTagsScorer("art"),
+  sport: createTagsScorer("sport"),
+  writing: createTagsScorer("writing"),
+} as const;
