+++ test/util.test.ts
@@ -1,7 +1,5 @@
 import { Immutable } from "@lauf/lauf-store";
-import { CATEGORIES } from "../src/data";
-import { SCORENAMES } from "../src/score";
-import { Entry } from "../src/types";
+import { CATEGORIES, Entry, SCORENAMES } from "../src/types";
 import { sortEntries } from "../src/util";
 
 /** Creates minimal entry with text fields varying by index, no tags, fixed dates
