+++ tsconfig.jest.json
@@ -1,0 +1,1 @@
+{ "extends": "./tsconfig.json", "compilerOptions": { "jsx": "react" } }
